+++ Pipfile
@@ -3,16 +3,18 @@ name = "pypi"
 url = "https://pypi.org/simple"
 verify_ssl = true
 
+[scripts]
+kiki = "python kiki run"
+
 [dev-packages]
 pylint = "*"
 flake8 = "*"
 
 [packages]
-discord-py = "~=1.4.1"
+discord.py = "~=1.4.1"
 click = "~=7.1.2"
 aioredis = "~=1.3.1"
-pyenchant = "~=3.1.1"
-pyyaml = "*"
+pyyaml = "~=5.3.1"
 
 [requires]
 python_version = "3.7"
