+++ manifests/kiki.yaml
@@ -24,6 +24,12 @@ spec:
             limits:
               cpu: "100m"
               memory: "100M"
+          readinessProbe:
+            exec:
+            - cat
+            - /tmp/healthy
+            initialDelaySeconds: 10
+            periodSeconds: 5
           env:
             - name: REDIS_URL
               value: "redis://redis"
